Add current alias to all files
HTML redirects are not put in place for directories that already have
content so this is effectively a no-op on the website. The benefit is
that once a page has an alias in place, it can be moved without worry
about breaking links.

It is still expected that writers update the aliases on file moves but
this is a small layer of protection against missing this step.

Signed-off-by: Jack Baldry <jack.baldry@grafana.com>